refactor(tests): extract columnar path-assertion log-capture harness to tests/common
Eight `columnar_*` path-assertion / parity test binaries each carried a
byte-for-byte-near-identical ~50-65 line copy of the process-global
`log`-capture harness (`CaptureLogger`, `LOG_BUFFER`, `buffer()`,
`impl Log`, `static LOGGER`, `init_logger()`, `take_logs()`), the
`SERIAL` env-var-toggling mutex (7 of 8), and the `execute_sql` /
`run_select` boilerplate. Consolidate all of it into the shared
`crates/vibesql-executor/tests/common/mod.rs` (included per-binary via
`mod common;`), so each test binary still compiles its OWN private copy
of the statics — the per-binary logger/env-var isolation the design
relies on is preserved.

No behavior change:
- The capture level varied per file (Info in 2, Debug in 6), so the
  extracted `init_logger(capture_upto: log::Level)` takes the level as a
  parameter and stores it in an atomic read by the shared `CaptureLogger`.
  Each call site passes the exact level its inlined copy used.
- `run_select` had two incompatible return types across the 8 files
  (`Vec<vibesql_storage::Row>` in 4, `Vec<Vec<SqlValue>>` in 4). Rather
  than change any call site's types, `common` exports both: `run_select`
  (rows) and `run_select_values` (projected values). The 4 value-returning
  files import `run_select_values as run_select`, so their bodies are
  untouched.
- `SERIAL` is `#[allow(dead_code)]` so the one binary that doesn't toggle
  env vars (is_null) simply doesn't import it.

All 8 binaries pass individually, concurrently, and under a full
`cargo test -p vibesql-executor` run; no other binary consuming
`tests/common/mod.rs` is affected (additions are additive). ~620 lines of
duplicated harness collapse into ~130 shared lines.

Closes #6010
