+++ crates/vibesql-executor/tests/columnar_join_stddev_variance_tests.rs
@@ -6,81 +6,17 @@
 //! (`VIBESQL_DISABLE_COLUMNAR_JOIN=1`) within a float epsilon. Uses the PR #6006
 //! SERIAL-mutex pattern because the env var and the log buffer are process-global.
 
-use std::sync::{Mutex, OnceLock};
-
-use log::{Level, LevelFilter, Log, Metadata, Record};
-use vibesql_executor::{CreateTableExecutor, InsertExecutor};
-use vibesql_parser::Parser;
+use log::Level;
 use vibesql_storage::Database;
 use vibesql_types::SqlValue;
 
-struct CaptureLogger;
-
-static LOG_BUFFER: OnceLock<Mutex<Vec<String>>> = OnceLock::new();
-
-fn buffer() -> &'static Mutex<Vec<String>> {
-    LOG_BUFFER.get_or_init(|| Mutex::new(Vec::new()))
-}
-
-impl Log for CaptureLogger {
-    fn enabled(&self, _metadata: &Metadata) -> bool {
-        true
-    }
-    fn log(&self, record: &Record) {
-        if record.level() <= Level::Debug {
-            buffer().lock().unwrap().push(format!("{}", record.args()));
-        }
-    }
-    fn flush(&self) {}
-}
-
-static LOGGER: CaptureLogger = CaptureLogger;
-
-fn init_logger() {
-    let _ = log::set_logger(&LOGGER);
-    log::set_max_level(LevelFilter::Debug);
-}
-
-fn take_logs() -> Vec<String> {
-    std::mem::take(&mut *buffer().lock().unwrap())
-}
-
-static SERIAL: Mutex<()> = Mutex::new(());
+mod common;
+use common::{execute_sql, init_logger, run_select_values as run_select, take_logs, SERIAL};
 
 /// Row-fallback marker: a joined GROUP BY key couldn't be resolved to a column.
 const ROW_FALLBACK: &str = "Columnar join: some GROUP BY columns couldn't be resolved";
 /// Positive marker emitted only when the columnar join path produced the result.
 const COLUMNAR_JOIN_SUCCEEDED: &str = "Columnar join execution succeeded";
-
-fn execute_sql(db: &mut Database, sql: &str) {
-    for sql_stmt in sql.split(';') {
-        let trimmed = sql_stmt.trim();
-        if trimmed.is_empty() {
-            continue;
-        }
-        let stmt = Parser::parse_sql(trimmed).expect("Failed to parse SQL");
-        match stmt {
-            vibesql_ast::Statement::CreateTable(s) => {
-                CreateTableExecutor::execute(&s, db).expect("CREATE TABLE failed");
-            }
-            vibesql_ast::Statement::Insert(s) => {
-                InsertExecutor::execute(db, &s).expect("INSERT failed");
-            }
-            other => panic!("Unsupported statement type: {:?}", other),
-        }
-    }
-}
-
-fn run_select(db: &Database, sql: &str) -> Vec<Vec<SqlValue>> {
-    let stmt = Parser::parse_sql(sql).expect("Failed to parse SELECT");
-    if let vibesql_ast::Statement::Select(select_stmt) = stmt {
-        let rows =
-            vibesql_executor::SelectExecutor::new(db).execute(&select_stmt).expect("SELECT failed");
-        rows.into_iter().map(|r| r.values.to_vec()).collect()
-    } else {
-        panic!("Expected SELECT");
-    }
-}
 
 fn as_f64(v: &SqlValue) -> Option<f64> {
     match v {
@@ -170,7 +106,7 @@ const SPELLINGS: [&str; 6] =
 #[test]
 fn join_group_by_stddev_variance_all_spellings_match_row_path() {
     let _guard = SERIAL.lock().unwrap_or_else(|e| e.into_inner());
-    init_logger();
+    init_logger(Level::Debug);
     let mut db = Database::new();
     setup_inner(&mut db, 800);
 
@@ -185,7 +121,7 @@ fn join_group_by_stddev_variance_all_spellings_match_row_path() {
 #[test]
 fn join_group_by_mixed_aggregates_match_row_path() {
     let _guard = SERIAL.lock().unwrap_or_else(|e| e.into_inner());
-    init_logger();
+    init_logger(Level::Debug);
     let mut db = Database::new();
     setup_inner(&mut db, 800);
 
@@ -200,7 +136,7 @@ fn join_group_by_mixed_aggregates_match_row_path() {
 #[test]
 fn join_group_by_stddev_variance_null_values_match_row_path() {
     let _guard = SERIAL.lock().unwrap_or_else(|e| e.into_inner());
-    init_logger();
+    init_logger(Level::Debug);
     let mut db = Database::new();
     execute_sql(&mut db, "CREATE TABLE l (k INTEGER, a INTEGER, v INTEGER)");
     execute_sql(&mut db, "CREATE TABLE r (k INTEGER, d INTEGER)");
